오류 수정 후 재업로드

+++ java/resproject/Review.java
@@ -1,6 +1,7 @@
 package resproject;
 
 import java.sql.PreparedStatement;
+import java.sql.Statement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -35,13 +36,23 @@ public class Review extends Conn{
 	public int Drop(int id) {
 		
 		PreparedStatement pstmt = null;
+		Statement stmt = null;
 		
 		try {
-			String sql = "DELETE FROM review WHERE id=?";
-			pstmt = conn.prepareStatement(sql);
-			pstmt.setInt(1, id);
+			String sql = "SELECT * FROM review WHERE id=" + id;
+			stmt = conn.createStatement();
+			ResultSet rs = stmt.executeQuery(sql);
 			
-			return pstmt.executeUpdate();
+			if(rs.next()) {
+				sql = "DELETE FROM review WHERE id=?";
+				pstmt = conn.prepareStatement(sql);
+				pstmt.setInt(1, id);
+				
+				return pstmt.executeUpdate();
+			} else {
+				System.out.println("\n아이디가 존재하지 않습니다.");
+				return -1;
+			}
 			
 		} catch(SQLException e) {
 			System.out.println(e.getMessage());
@@ -53,17 +64,27 @@ public class Review extends Conn{
 	public int Update(String user_id, int id, double rate, String comment) {
 		
 		PreparedStatement pstmt = null;
+		Statement stmt = null;
 		
 		try {
-			String sql = "UPDATE review SET rate=?, comment=? "
-					+ "WHERE user_id=? AND id=?";
-			pstmt = conn.prepareStatement(sql);
-			pstmt.setDouble(1, rate);
-			pstmt.setString(2, comment);
-			pstmt.setString(3, user_id);
-			pstmt.setInt(4, id);
+			String sql = "SELECT * FROM review WHERE id=" + id;
+			stmt = conn.createStatement();
+			ResultSet rs = stmt.executeQuery(sql);
 			
-			return pstmt.executeUpdate();
+			if(rs.next()) {
+				sql = "UPDATE review SET rate=?, comment=? "
+						+ "WHERE user_id=? AND id=?";
+				pstmt = conn.prepareStatement(sql);
+				pstmt.setDouble(1, rate);
+				pstmt.setString(2, comment);
+				pstmt.setString(3, user_id);
+				pstmt.setInt(4, id);
+				
+				return pstmt.executeUpdate();
+			} else {
+				System.out.println("\n아이디가 존재하지 않습니다.");
+				return -1;
+			}
 			
 		} catch(SQLException e) {
 			System.out.println(e.getMessage());
